Api en azure funcionando
Sumado el script para Azure, el manual de usuario y muchos cambios.
Mejorados los metodos de la API más que nada

+++ TrabajoFinalPrueba1/PegasusV1/PegasusV1/Controllers/EventoController.cs
@@ -39,33 +39,32 @@ namespace PegasusV1.Controllers
 
         [HttpGet]
         [Route("GetById")]
-        public async Task<Evento> GetById(int id)
+        public async Task<Evento?> GetById(int id)
         {
             return await EventoService.GetById(id);
         }
 
         [HttpPost]
         [Route("CreateEvento")]
-        public async Task<Evento> CreateEvento(string eventos)
+        public async Task<Evento> CreateEvento(Evento evento)
         {
-            Evento evento = JsonConvert.DeserializeObject<Evento>(eventos);
             return await EventoService.Create(evento);
         }
 
-        [HttpPost]
+        [HttpPut]
         [Route("UpdateEvento")]
-        public async Task<Evento> UpdateEvento(string eventos)
+        public async Task<Evento> UpdateEvento(Evento evento)
         {
-            Evento evento = JsonConvert.DeserializeObject<Evento>(eventos);
             return await EventoService.Update(evento);
         }
 
-        [HttpPost]
+        [HttpDelete]
         [Route("DeleteEvento")]
-        public async void DeleteEvento(string eventos)
+        public async Task DeleteEvento(int id)
         {
-            Evento evento = JsonConvert.DeserializeObject<Evento>(eventos);
-            EventoService.Delete(evento);
+            Evento? evento = await GetById(id);
+            if (evento != null) 
+                await EventoService.Delete(evento);
         }
     }
 }
